requirements.txt
zeroconf
hkdf
ed25519
dbus-python
pycairo
PyGObject
gatt
cryptography>=2.5
coverage
flake8